separate blake_context
